+++ migrations/env.py
@@ -21,6 +21,7 @@ from geoalchemy2 import alembic_helpers
 
 import geometrikks.domain.geo.models  # noqa: F401
 import geometrikks.domain.logs.models  # noqa: F401
+import geometrikks.domain.imports.models  # noqa: F401
 # --- End local additions ------------------------------------------------------
 
 if TYPE_CHECKING: